backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.6.1
python-multipart==0.0.9
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-core==0.3.15
google-genai==1.1.0
deepgram-sdk==3.8.3
youtube-transcript-api==0.6.3
pypdf2==3.0.1
pillow==11.0.0
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.0

